+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Remove `EveOnline::Bookmark`
 * Remove `EveOnline::BookmarkFolder`
 * Remove `EveOnline::XML::CharacterBookmarks`. Use `EveOnline::ESI::CharacterBookmarks` instead.
+* Add `EveOnline::ESI::Models::BookmarkFolder`
 
 **v0.14.0**